+++ x/ecocredit/basket/types/v1/tx.pb.go
@@ -490,8 +490,8 @@ func (m *MsgTakeResponse) GetCredits() []*BasketCredit {
 type MsgUpdateBasketFee struct {
 	// authority is the address of the governance account.
 	Authority string `protobuf:"bytes,1,opt,name=authority,proto3" json:"authority,omitempty"`
-	// basket_fee is the basket creation fee. If not set, the basket creation fee
-	// will be removed and no fee will be required to create a basket.
+	// fee is the basket creation fee. If not set, the basket creation fee will be
+	// removed and no fee will be required to create a basket.
 	Fee *types.Coin `protobuf:"bytes,2,opt,name=fee,proto3" json:"fee,omitempty"`
 }
 
@@ -671,8 +671,9 @@ type MsgClient interface {
 	// Take takes credits from a basket starting from the oldest
 	// credits first.
 	Take(ctx context.Context, in *MsgTake, opts ...grpc.CallOption) (*MsgTakeResponse, error)
-	// UpdateBasketFee is a governance method that allows for updating of fees to
-	// be used for the basket creation fee.
+	// UpdateBasketFee is a governance method that allows for updating the basket
+	// creation fee. If not set, the basket creation fee will be removed and no
+	// fee will be required to create a basket.
 	//
 	// Since Revision 2
 	UpdateBasketFee(ctx context.Context, in *MsgUpdateBasketFee, opts ...grpc.CallOption) (*MsgUpdateBasketFeeResponse, error)
@@ -731,8 +732,9 @@ type MsgServer interface {
 	// Take takes credits from a basket starting from the oldest
 	// credits first.
 	Take(context.Context, *MsgTake) (*MsgTakeResponse, error)
-	// UpdateBasketFee is a governance method that allows for updating of fees to
-	// be used for the basket creation fee.
+	// UpdateBasketFee is a governance method that allows for updating the basket
+	// creation fee. If not set, the basket creation fee will be removed and no
+	// fee will be required to create a basket.
 	//
 	// Since Revision 2
 	UpdateBasketFee(context.Context, *MsgUpdateBasketFee) (*MsgUpdateBasketFeeResponse, error)